requirements.txt
alembic==1.13.1
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
requests==2.32.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.2
